# Reuse Snappy decoder, for performance