Part of #624 

Calling `Reset()` rather than discarding and creating a new one allows buffers to be re-used, which reduces garbage-collection.

The benefit comes when you decode more than one chunk in a row - 55%-60% less time in these benchmarks:

before:
```
BenchmarkDecode1-2       	  200000	    102466 ns/op	  155771 B/op	     117 allocs/op
BenchmarkDecode100-2     	    2000	  10742609 ns/op	15577124 B/op	   11632 allocs/op
BenchmarkDecode10000-2   	      20	 829417236 ns/op	1557639768 B/op	 1163165 allocs/op
```

after:
```
BenchmarkDecode1-2       	  200000	    113408 ns/op	  155772 B/op	     117 allocs/op
BenchmarkDecode100-2     	    3000	   4202005 ns/op	  967774 B/op	   11335 allocs/op
BenchmarkDecode10000-2   	      30	 378176958 ns/op	82100856 B/op	 1133154 allocs/op
```